Desistência do sqlite e adoção do banco para os dados estaticos como o sqlserver

+++ Dominio/src/Models/Logradouro.cs
@@ -6,7 +6,7 @@ namespace Rio.SMF.CCU.Ouvidoria.Dominio.Models
     public partial class Logradouro
     {
 
-        public long LogradouroId { get; set; }
+        public int LogradouroId { get; set; }
         public string IdBairro { get; set; }
         public string Nome { get; set; }
